TST: move SWT tests from test_multilevel.py into a separate test_swt.py
